refactor(user): moved attendee, organizer, and referral service into one under user service

+++ src/main/java/com/dti/ecim/user/entity/Attendee.java
@@ -47,5 +47,5 @@ public class Attendee {
 
     @PositiveOrZero
     @Column(name = "points")
-    private Long points;
+    private int points;
 }
